Fix uint8_t inconsistencies.

+++ CCS/wisp-base/RFID/rfid.h
@@ -27,7 +27,7 @@
 typedef struct {
 	uint8_t* epcBuf;
 	uint16_t* writeBufPtr;
-	uint8_t* blockWriteBufPtr;
+	uint16_t* blockWriteBufPtr;
 	uint16_t* blockWriteSizePtr;
 	uint8_t* readBufPtr;
 } WISP_dataStructInterface_t;
